Allow use outside of browser environment
Stub in browser globals if they are not set
