+++ standalone/index.js
@@ -1,28 +1,5 @@
-const WordFilter = require('./WordFilter.js');
-const Styles = require('./Styles.js');
-
-StylesTool = Styles();
-
-/*
-    create a mock editor object, expected by plugin;
-    must include Styles de/serializers
-*/
-const editor = {
-  settings: {},
-  dom: {
-    parseStyle: StylesTool.parse,
-    serializeStyle: StylesTool.serialize,
-  },
-};
-
-/*
-    also mock the param function required by editor
-*/
-editor.getParam = function(key, defaultValue) {
-  const value = key in editor.settings ? editor.settings[key] : defaultValue;
-
-  return value;
-};
+const editor = require("./mocks.js");
+const WordFilter = require("./WordFilter.js");
 
 module.exports = function CleanWordHTML(content, settings) {
   editor.settings = settings;
